+++ demo/src/main.ts
@@ -5,5 +5,6 @@ import '../../src/assets/scss/index.scss';
 import '../../src/assets/scss/themes/monokai-dark.scss';
 import '../../src/assets/scss/themes/visual-studio-light.scss';
 import '../../src/assets/scss/themes/atom-dark.scss';
+import '../../src/assets/scss/themes/atom-light.scss';
 
 createApp(App).mount('#app');